# api-gateway/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Learning Platform API
  description: API для управления образовательной платформой
  version: '1.0'
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
...
  /module/import:
    post:
      tags:
        - Module Controller
      summary: Импорт модуля из Markdown файла
      description: Позволяет создать новый модуль из .md файла. Требуются права ADMIN или MENTOR
      operationId: importModuleFromMarkdown
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
                request:
                  $ref: '#/components/schemas/CreateModuleRequest'
        required: true
      responses:
        '200':
          description: Информация о модуле
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ModuleDto'
        '400':
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                type: object
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                type: object
  /module/create:
    post:
      tags:
        - Module Controller
      summary: Создать модуль
      description: Позволяет создать новый модуль. Требуются права ADMIN или MENTOR
      operationId: createModule
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateModuleRequest'
        required: true
      responses:
        '200':
          description: Модуль успешно создан
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ModuleDto'
        '400':
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                type: object
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                type: object
  /course/create:
    post:
      tags:
        - Course Controller
      summary: Создать курс
      description: Позволяет создать новый курс. Требуются права ADMIN или MENTOR
      operationId: createCourse
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCourseRequest'
        required: true
      responses:
        '200':
          description: Курс успешно создан
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CourseDto'
        '400':
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CourseDto'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CourseDto'
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CourseDto'
...
  /module/{courseId}/{moduleId}:
    get:
      tags:
        - Module Controller
      summary: Получить модуль по ID
      description: Возвращает информацию о модуле по его идентификатору
      operationId: getModuleById
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация о модуле
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ModuleDto'
        '400':
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ModuleDto'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ModuleDto'
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ModuleDto'
    delete:
      tags:
        - Module Controller
      summary: Удалить модуль
      description: Позволяет удалить модуль. Требуются права ADMIN или MENTOR
      operationId: deleteModule
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Модуль успешно удален
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ModuleDto'
        '400':
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                type: object
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                type: object
  /course/{courseId}:
    get:
      tags:
        - Course Controller
      summary: Получить курс по ID
      description: Возвращает информацию о курсе по его идентификатору
      operationId: getCourseById
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация о курсе
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CourseDto'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CourseDto'
        '404':
          description: Курс не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CourseDto'
    delete:
      tags:
        - Course Controller
      summary: Удалить курс
      description: Позволяет удалить курс. Требуются права ADMIN или MENTOR
      operationId: deleteCourse
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Курс успешно удален
          content:
            '*/*':
              schema:
                type: object
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                type: object
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                type: object
        '404':
          description: Курс не найден
          content:
            '*/*':
              schema:
                type: object
  /course/all:
    get:
      tags:
        - Course Controller
      summary: Получить все курсы
      description: Возвращает список всех курсов, включая неактивные
      operationId: getAllCourses
      responses:
        '200':
          description: Список всех курсов
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseDto'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseDto'
  /course/all/active:
    get:
      tags:
        - Course Controller
      summary: Получить активные курсы
      description: Возвращает список всех активных курсов
      operationId: getAllActiveCourses
      responses:
        '200':
          description: Список активных курсов
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseDto'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseDto'
...
  /admin/course/{courseId}:
    get:
      tags:
        - Admin Course Controller
      summary: Получить курс по ID
      description: Возвращает информацию о курсе по его идентификатору
      operationId: adminGetCourseById
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация о курсе
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CourseDto'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CourseDto'
        '404':
          description: Курс не найден
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CourseDto'
  /admin/course/all:
    get:
      tags:
        - Admin Course Controller
      summary: Получить все курсы
      description: Возвращает список всех курсов, включая неактивные
      operationId: adminGetAllCourses
      parameters:
        - name: pageNumber
          in: query
          required: true
          schema:
            type: integer
            format: int32
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Список всех курсов
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseDto'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageCourseDto'
components:
  schemas:
...
    ModuleDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        moduleTitle:
          type: string
        moduleOrderNumber:
          type: integer
          format: int32
        moduleContent:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateModuleRequest:
      required:
        - courseId
        - moduleContentDescription
        - moduleOrderNumber
        - moduleTitle
      type: object
      properties:
        courseId:
          type: integer
          description: ID курса
          format: int64
          example: 123
        moduleTitle:
          type: string
          description: Название модуля
          example: 'Модуль 1: Основы Spring Boot'
        moduleOrderNumber:
          type: integer
          description: Порядковый номер модуля в курсе
          format: int32
          example: 1
        moduleContentDescription:
          type: string
          description: Содержимое модуля в формате Markdown
          example: |-
            # Введение

            Этот модуль содержит основы Spring Boot...
      description: Запрос на создание модуля курса
    CourseDto:
      type: object
      properties:
        id:
          type: integer
          description: ID курса
          format: int64
          example: 1
        courseTitle:
          type: string
          description: Название курса
          example: Основы Spring Boot
        courseDescription:
          type: string
          description: Описание курса
          example: Курс по основам Spring Framework
        isActive:
          type: boolean
          description: Флаг активности курса
          example: true
        createdAt:
          type: string
          description: Дата создания курса
          format: date-time
        author:
          $ref: '#/components/schemas/UserInfoDto'
        modules:
          type: array
          description: Список модулей курса
          items:
            $ref: '#/components/schemas/ModuleDto'
      description: Информация о курсе
    UserInfoDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        role:
          type: string
          enum:
            - USER
            - ADMIN
            - MENTOR
        firstName:
          type: string
        lastName:
          type: string
        tgNickname:
          type: string
        tgChatId:
          type: integer
          format: int64
      description: ID автора курса
      example: 5
    CreateCourseRequest:
      required:
        - courseDescription
        - courseName
      type: object
      properties:
        courseName:
          type: string
          description: Название курса
          example: Основы Spring Boot
        courseDescription:
          type: string
          description: Описание курса
          example: Курс по основам Spring Framework
      description: Запрос на создание курса
...
    PageableObject:
      type: object
      properties:
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        direction:
          type: string
        nullHandling:
          type: string
        ascending:
          type: boolean
        property:
          type: string
        ignoreCase:
          type: boolean
...
    PageCourseDto:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/CourseDto'
        number:
          type: integer
          format: int32
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean